Fixed issues with constants.ts and improve code styling
