+++ packages/app/components/Button.tsx
@@ -1,13 +1,7 @@
 import React from 'react';
-import { Button as RNButton, NativeSyntheticEvent, NativeTouchEvent, ColorValue } from 'react-native';
+import { ButtonProps, Button as RNButton } from 'react-native';
 
-type ButtonProps = {
-  title: string;
-  onPress: (ev: NativeSyntheticEvent<NativeTouchEvent>) => void;
-  color?: ColorValue | undefined;
-  accessibilityLabel?: string | undefined;
-  disabled?: boolean | undefined;
-}
+
 
 
 function Button(props: ButtonProps) {
