+++ packages/app/components/Screen.tsx
@@ -1,9 +1,8 @@
-import React, { ReactNode } from 'react';
+import { ScreenProps } from '@app/types/Screen';
+import React from 'react';
 import { SafeAreaView, StatusBar } from 'react-native';
 
-type ScreenProps = {
-  children: ReactNode
-}
+
 
 function Screen(props: ScreenProps ) {
   const { children, ...rest } = props;
