+++ packages/app/package.json
@@ -1,13 +1,13 @@
 {
   "name": "app",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "index.tsx",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "dev": "ttab expo start",
+    "dev": "expo start",
     "test": "jest"
   },
   "dependencies": {
