+++ packages/app/tsconfig.json
@@ -27,7 +27,7 @@
   "exclude": [
     "node_modules",
     "babel.config.js",
-    "metro.config.js",
+    "metro.config.ts",
     "jest.config.js"
   ],
   "extends": "expo/tsconfig.base"
