+++ packages/app/types/Button.ts
@@ -1,0 +1,9 @@
+import { ColorValue, NativeSyntheticEvent, NativeTouchEvent } from "react-native";
+
+export type ButtonProps = {
+  title: string;
+  onPress: (ev: NativeSyntheticEvent<NativeTouchEvent>) => void;
+  color?: ColorValue | undefined;
+  accessibilityLabel?: string | undefined;
+  disabled?: boolean | undefined;
+}
