+++ packages/app/types/Screen.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react"
+
+export type ScreenProps = {
+  children: ReactNode
+}
