+++ packages/app/utils/constants.ts
@@ -1,0 +1,3 @@
+const URL = "https://af57-2620-df-8000-4707-0-2-7f9e-5884.ngrok.io";
+
+export const ENDPOINT = `${URL}/api`;
