--- packages/app/utils/constants.tsx
@@ -1,3 +1,0 @@
-const URL = "https://fc85-2620-df-8000-4707-0-2-1ab7-9e63.ngrok.io";
-
-export const ENDPOINT = `${URL}/api`;
